Converted class to trait
No longer using validation exception factories; using throwing methods instead,
which fixes #11.

Also, now throwing `ValidationExceptionInterface` for errors *during* validation.
